ground_station/archive/requirements.txt
numpy
matplotlib
pyserial
contextily
requests
